Adds Controller service, tests, Response body and Request body

+++ src/main/java/com/unit_conversion/service/ConvertService.java
@@ -2,7 +2,7 @@ package com.unit_conversion.service;
 
 public interface ConvertService {
 
-    double convert(double valueToConvert);
+    float convert(float valueToConvert);
 
 
 }
